requirements.txt
keyboard~=0.13.5
PyAutoGUI~=0.9.54
googletrans~=3.1.0a0
winclip32~=0.6.0
infi.systray~=0.1.12